feat(bot): email and tg_nick dual validation

+++ pkg/bot/remove.go
@@ -1,6 +1,10 @@
 package bot
 
-import tb "gopkg.in/tucnak/telebot.v2"
+import (
+	"database/sql"
+
+	tb "gopkg.in/tucnak/telebot.v2"
+)
 
 func (s *Sylon) registerRemoveHandler() {
 	s.B.Handle("/remove", func(m *tb.Message) {
@@ -9,25 +13,33 @@ func (s *Sylon) registerRemoveHandler() {
 			return
 		}
 
-		var ownerChatID int64
+		var ownerUsername sql.NullString
 		if err := s.R.Get(
-			&ownerChatID,
+			&ownerUsername,
 			`--sql
-			select personal_chat
-			from client inner join project
-				on project.client_id = client.id
-			where project.id = $1`,
+				select tg_nick
+				from client inner join project
+					on project.client_id = client.id
+				where project.id = $1`,
 			project.ID,
 		); err != nil {
 			s.JustSend(m.Chat, "Failed to verify owner of the project.")
+			return
 		}
 
-		if m.Sender.ID != ownerChatID {
+		if !ownerUsername.Valid {
+			s.JustSend(m.Chat, "Failed to verify owner of the project.")
+			return
+		}
+
+		if m.Sender.Username != ownerUsername.String {
 			s.JustSend(m.Chat, "You are not the owner of this project.")
+			return
 		}
 
 		if _, err := s.R.Exec("delete from project where id = $1", project.ID); err != nil {
 			s.JustSend(m.Chat, "Failed to remove project.")
+			return
 		}
 		s.JustSend(m.Chat, "Project removed.")
 	})
